Moved loadSpecification helper function
